NEW Restructured queue referencing in job
Moved the retrieval of the URLs to the 'setup' method of the job to ensure that
the job processes all URLs requested in the last 5 minutes, so that the
processing of URLs doesn't get delayed by 5 minutes while the job waits to be
executed. Means the count that appears in the queue doesn't show up until
runtime, but that's a reasonable concession.